requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
boto3==1.34.0
pdfplumber==0.10.3
Pillow==10.1.0
requests==2.31.0
